+++ patientsearch/src/js/components/patientList/PatientList.js
@@ -15,7 +15,7 @@ import Pagination from "./Pagination";
 import OverlayElement from "../OverlayElement";
 import TestPatientsCheckbox from "./TestPatientsCheckbox";
 import * as constants from "../../constants/consts";
-import { addMamotoTracking, getUrlParameter } from "../../helpers/utility";
+import { addMamotoTracking, getUrlParameter, hasFlagForCheckbox } from "../../helpers/utility";
 
 export default function PatientListTable() {
   const patientListCtx = usePatientListContext();
@@ -35,13 +35,6 @@ export default function PatientListTable() {
     openLoadingModal,
     setOpenLoadingModal
   } = usePatientListContext();
-
-  const hasFlagForCheckbox = (flagId) => {
-    const flagQueryString = getUrlParameter("flags");
-    if (!flagQueryString) return false;
-    if (!flagId) return false;
-    return flagQueryString === flagId;
-  };
 
   const renderTitle = () => {
     const title = appSettings["SEARCH_TITLE_TEXT"]
@@ -73,7 +66,7 @@ export default function PatientListTable() {
     return (
       <MyPatientsCheckbox
         label={getAppSettingByKey("MY_PATIENTS_FILTER_LABEL")}
-        checked={hasFlagForCheckbox("following")}
+        checked={hasFlagForCheckbox(constants.FOLLOWING_FLAG)}
       ></MyPatientsCheckbox>
     );
   };
