+++ patientsearch/src/js/constants/consts.js
@@ -133,3 +133,4 @@ export const MORE_MENU_KEY = "MORE_MENU";
 export const ACCESS_TOKEN_KEY = "access_token";
 export const REALM_ACCESS_TOKEN_KEY = "realm_access";
 export const MAX_MAIN_TABLE_WIDTH = "1280px";
+export const FOLLOWING_FLAG = "following";
